pyproject.toml
[project]
name="redis_learning"
version="0.1"
requires-python=">=3.11"
dependencies=[
    "fastapi[standard]",
    "uvicorn[standard]",
    "fastapi-redis-session",
    "black>=25.9.0",
    "python-dotenv>=1.1.1",
]
